completing the app

+++ app/src/main/java/com/thechance/season2week10/models/Condition.kt
@@ -6,6 +6,6 @@ import com.google.gson.annotations.SerializedName
 data class Condition(
     @SerializedName("text") var text: String? = null,
     @SerializedName("icon") var icon: String? = null,
-    @SerializedName("code") var code: Int? = null
+    @SerializedName("code") var code: Double? = null
 
 )
